Updated license and TODO, massive MJPG improvements
MJPG now has a massively improved command. Instead of transcoding the
video for no reason, instead we ask the camera driver for MJPG data so
that we don't have to transcode raw video if we don't have to.

MJPG's resolution and FPS can now also theoretically be changed from a
config file. The frontend components for this however are not
implemented.

Updated notice in first time setup since we're not in alpha anymore.