Changed config program to TP-Config, added Octo-Settings
Changed "octo-config" to "tp-config" as "octo-config" was misleading.

Added octo-settings, a command line utility that can read and set
OctoPrint settings.

Added OctoPrint config menu, in which I added the ability to have
localhost autologin as a user in OctoPrint

Improved grammar and verbosity in some text dialogs

Added message to the end of first-time explaining how to configure
TouchPrint

Updated README and TODO